+++ redux/basic_types.py
@@ -117,7 +117,9 @@ class CreateStoreOptions(Immutable, Generic[Action, Event]):
 
 class AutorunOptions(Immutable, Generic[AutorunOriginalReturnType]):
     default_value: AutorunOriginalReturnType | None = None
-    initial_run: bool = True
+    initial_call: bool = True
+    auto_call: bool = True
+    reactive: bool = True
     keep_ref: bool = True
     subscribers_initial_run: bool = True
     subscribers_keep_ref: bool = True