+++ redux/serialization_mixin.py
@@ -29,7 +29,7 @@ class SerializationMixin:
             return [cls.serialize_value(i) for i in obj]
         if is_immutable(obj):
             return cls._serialize_dataclass_to_dict(obj)
-        msg = f'Unable to serialize object with type `{type(obj)}`.'
+        msg = f'Unable to serialize object with type `{type(obj)}`'
         raise TypeError(msg)
 
     @classmethod